pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-agent-system"
version = "0.1.0"
description = "Terminal-first autonomous agent with tools, planning, and web dashboard"
readme = "README.md"
requires-python = ">=3.9"
authors = [{name = "Agent System"}]
license = {text = "MIT"}
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.12.0",
  "aiofiles>=23.0.0",
  "openai>=1.0.0",
  "anthropic>=0.7.0",
  "chromadb>=0.4.0",
  "cryptography>=41.0.0",
  "python-jose[cryptography]>=3.3.0",
  "structlog>=23.0.0",
  "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "black>=23.0.0",
  "ruff>=0.1.0",
  "mypy>=1.6.0",
  "coverage>=7.3.0",
]

[project.scripts]
agent-chat = "agent_system.chat_cli:main"
agent-term = "agent_system.terminal_cli:main"
agent-run = "agent_system.__main__:main"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]
ignore = ["E203", "E266", "E501", "W503"]

[tool.ruff.isort]
known-first-party = ["agent_system"]

[tool.pytest.ini_options]
addopts = "-q"

[tool.coverage.run]
branch = true
source = ["agent_system"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
ignore_missing_imports = true